feat(auth): added service that handles authentication
This service handles authentication with the desec token based authentication.
For login, email and password are POSTed to /api/auth/login, a auth_token is
expected to be sent back to the client. The auth_token will then be set as a
header value 'Authentication: Token <auth_token>'. For logout, a request will
be POSTed to /api/auth/logout and, if successful, the header information will
be deleted.