feat(dyndns): signup form for the dynDNS product
The new form asks for email address and a domain name below of dedyn.io,
like nils.dedyn.io. After submitting, the app attempts to register an user
account with the given email address. If successful, login is made and
the registration of the domain is attempted. If successful, the user sees
a confirmation message and is offered to click the "account setup" button.
This button is not yet linked to any page!

If user account registration fails, the user is offered to use a different
email address.

If domain registration fails, the user remains logged in and is offered to
use a different domain name.

The user registration uses a random password that is concealed from the user
at the time of this commit. Closing the browser windows after successful
user registration, but before successful domain registration renderes the
user account unusable. This ought to be fixed eventually! However, for dynDNS
updates the user can use the API auth code that is sent by email.