BREAKING CHANGE - DO NOT DEPLOY fix(grunt): remove unsupported steps
CSS compilation is not supported anymore the way we do it. This
commit removes the step from Grunt.

The resulting web site will have incomplete CSS. The purpose of
the commit is just to stop the build from breaking, until we
replace the website itself (which is on the way).